Fix build to allow `make ensure && make` work
It was currently expecting a `yarn.lock` in the root dir, but now
that we have moved the source under `./src`, this no longer works
(unless you had an old `yarn.lock` file sticking around in the root dir).
Update to just use a sub project Makefile, like we do elsewhere.